+++ src/builder.rs
@@ -5,7 +5,11 @@ use std::{
     fs::create_dir_all,
     path::Path,
     process::{Child, Command},
+    thread,
+    time::Duration,
 };
+
+use termal::{printc, printcln};
 
 use crate::{
     config::Config,
@@ -15,6 +19,8 @@ use crate::{
 };
 
 pub struct Builder {
+    /// Max number of threads running at the same time
+    pub thread_count: usize,
     /// C compiler
     pub cc: String,
     /// Linker
@@ -65,6 +71,8 @@ impl Builder {
         }
 
         Self {
+            thread_count: std::thread::available_parallelism()
+                .map_or(1, |t| t.get().checked_sub(2).unwrap_or(1)),
             cc: cc
                 .map(|c| c.to_owned())
                 .or(env::var("CC").ok())
@@ -83,15 +91,36 @@ impl Builder {
         let mut dep_deps = HashMap::new();
         let deps = get_dependencies(dir, &mut dep_deps)?;
 
-        for file in &deps {
-            if file.is_up_to_date()? {
-                continue;
-            }
+        let mut to_build = vec![];
 
-            self.sync_build_file(
-                file.direct.iter().map(|i| i.as_ref()),
-                file.file.as_ref(),
-            )?;
+        for file in deps {
+            if !file.is_up_to_date()? {
+                to_build.push(file);
+            }
+        }
+
+        if self.thread_count <= 1 {
+            for file in to_build {
+                self.sync_build_file(
+                    file.direct.iter().map(|i| i.as_ref()),
+                    file.file.as_ref(),
+                )?;
+            }
+        } else {
+            let mut threads = vec![];
+            if let Err(e) = self.parallel_build(to_build, &mut threads) {
+                for mut thread in threads {
+                    if let Err(e) = thread.wait() {
+                        printcln!(
+                            "      {'r bold}Error{'_}\
+                            failed to wait for thread: {}",
+                            e
+                        );
+                        _ = thread.kill();
+                    }
+                }
+                return Err(e);
+            }
         }
 
         let bin_dep = Dependency {
@@ -113,6 +142,51 @@ impl Builder {
 //===========================================================================//
 
 impl Builder {
+    fn parallel_build(
+        &self,
+        files: Vec<Dependency>,
+        threads: &mut Vec<Child>,
+    ) -> Result<()> {
+        const TIMEOUT: Duration = Duration::from_millis(10);
+
+        let mut files = files.into_iter();
+
+        while let Some(file) = files.next() {
+            threads.push(self.start_build_file(
+                file.direct.iter().map(|i| i.as_ref()),
+                &file.file,
+            )?);
+            if threads.len() == self.thread_count {
+                break;
+            }
+        }
+
+        'files: while let Some(file) = files.next() {
+            for t in threads.iter_mut() {
+                if let Some(res) = t.try_wait()? {
+                    if !res.success() {
+                        return Err(Error::ProcessFailed(res.code()));
+                    }
+                    *t = self.start_build_file(
+                        file.direct.iter().map(|i| i.as_ref()),
+                        &file.file,
+                    )?;
+                    continue 'files;
+                }
+            }
+            thread::sleep(TIMEOUT);
+        }
+
+        for thread in threads.into_iter() {
+            let res = thread.wait()?;
+            if !res.success() {
+                return Err(Error::ProcessFailed(res.code()));
+            }
+        }
+
+        Ok(())
+    }
+
     fn start_build_file<'a, S, I>(
         &self,
         input: I,
@@ -120,7 +194,7 @@ impl Builder {
     ) -> Result<Child>
     where
         S: AsRef<OsStr>,
-        I: Iterator<Item = S>,
+        I: Iterator<Item = S> + Clone,
     {
         if let Some(d) = output.parent() {
             create_dir_all(d)?;
@@ -131,10 +205,14 @@ impl Builder {
             .arg("-o")
             .arg(output)
             .arg("-c")
-            .args(input);
+            .args(input.clone());
 
         if self.print_command {
-            println!("{cmd:?}");
+            printc!("{'g bold}  Compiling{'_}");
+            for f in input {
+                print!(" {}", f.as_ref().to_string_lossy());
+            }
+            println!();
         }
         Ok(cmd.spawn()?)
     }
@@ -142,7 +220,7 @@ impl Builder {
     fn sync_build_file<'a, S, I>(&self, input: I, output: &Path) -> Result<()>
     where
         S: AsRef<OsStr>,
-        I: Iterator<Item = S>,
+        I: Iterator<Item = S> + Clone,
     {
         let res = self.start_build_file(input, output)?.wait()?;
         if !res.success() {
@@ -172,7 +250,7 @@ impl Builder {
             .args(input);
 
         if self.print_command {
-            println!("{cmd:?}");
+            printcln!("{'g bold}    Linking{'_} {}", output.to_string_lossy());
         }
 
         Ok(cmd.spawn()?)